Refactors Gradio Frontend into multiple Classes
