+++ sql_migration_assistant/main.py
@@ -1,0 +1,13 @@
+
+from sql_migration_assistant.frontend.GradioFrontend import GradioFrontend
+import os
+
+def main():
+    frontend = GradioFrontend()
+    frontend.app.queue().launch(
+        server_name=os.getenv("GRADIO_SERVER_NAME", "localhost"),
+        server_port=int(os.getenv("GRADIO_SERVER_PORT", 3001)),
+    )
+
+if __name__ == "__main__":
+    main()
